requirements.txt
# Django core
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Django utilities
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3

# Database
psycopg2-binary>=2.9.9         # PostgreSQL adapter (for production)

# Environment variables
python-dotenv>=1.0.0            # Load .env files

# Authentication (Phase 2)
# djangorestframework-simplejwt>=5.3.0  # JWT auth (uncomment when needed)

# Cloud storage for images (Cloudflare R2)
boto3>=1.34.0                   # S3-compatible API client
django-storages>=1.14.0         # Django storage backend for S3/R2

# Image processing
Pillow>=10.0.0                  # Image resize/compress for character portraits

# HTTP client for Patreon API
requests>=2.31.0                # HTTP requests library

# Production server
gunicorn>=21.2.0                # WSGI HTTP server for production
whitenoise>=6.6.0               # Static file serving for production
